Some changes in migrations to check before creating them

+++ database/migrations/2022_05_01_111159_alter_user_table_verify_token.php
@@ -13,14 +13,16 @@ class AlterUserTableVerifyToken extends Migration
      */
     public function up()
     {
-        Schema::create('activation_codes', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('user_id');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('token');
-            $table->unique(['user_id', 'token']);
-            $table->timestamp('expired_at');
-        });
+        if ( !Schema::hasTable('activation_codes') ) {
+            Schema::create('activation_codes', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->unsignedBigInteger('user_id');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->integer('token');
+                $table->unique(['user_id', 'token']);
+                $table->timestamp('expired_at');
+            });
+        }
     }
 
     /**
